Crate parquet_geospatial

Implementation of Geometry and Geography Encoding from Apache Parquet.

🚧 Work In Progress

This crate is under active development and is not yet ready for production use. If you are interested in helping, you can find more information on the GitHub Geometry issue

Modules

bounding

interval

testing

Testing utilities for geospatial Parquet types