parquet_geospatial::types

Struct Metadata

pub struct Metadata {
    pub crs: Option<Value>,
    pub algorithm: Option<Edges>,
}

The metadata associated with a WkbType.

Fields
crs: Option<Value>

The Coordinate Reference System (CRS) of the WkbType, if present.

This may be a raw string value (e.g., โEPSG:3857โ) or a JSON object (e.g., PROJJSON). Note: Common lon/lat CRS representations (EPSG:4326, OGC:CRS84) are canonicalized to None during serialization to match Parquet conventions.

algorithm: Option<Edges>

The edge interpolation algorithm of the WkbType, if present.

Implementations

impl Metadata

pub fn new(crs: Option<&str>, algorithm: Option<Edges>) -> Self

Constructs a new Metadata with the given CRS and algorithm.

If a CRS is provided, and can be parsed as JSON, it will be stored as a JSON object instead of its string representation.

pub fn type_hint(&self) -> Hint

Returns a Hint to the likely underlying Logical Type that this Metadata represents.

fn crs_is_lon_lat(&self) -> bool

Detect if the CRS is a common representation of lon/lat on the standard WGS84 ellipsoid