Crate parquet_geospatial

Implementation of Geometry and Geography Encoding from Apache Parquet.

🚧 Work In Progress

This crate is under active development and is not yet ready for production use. If you are interested in helping, you can find more information on the GitHub Geometry issue

Modules

bounding

interval

testing

Testing utilities for geospatial Parquet types

types 🔒

Structs

WkbMetadata

The metadata associated with a WkbType.

WkbType

Well-Known Binary (WKB) [ExtensionType] for geospatial data.

Enums

WkbEdges

The edge interpolation algorithms used with GEOMETRY logical types.

WkbTypeHint

Hints at the likely Parquet geospatial logical type represented by a Metadata.